+++ src/content/issue.js
@@ -33,7 +33,7 @@ function scrapeIssueData() {
 	}
 	for (const $report of $content.querySelectorAll('.row .row .col')) {
 		const report={
-			wasRead:$report.parentElement.classList.contains('text-muted'),
+			wasRead:$report.parentElement.parentElement.classList.contains('text-muted'),
 			lead:[],
 			text:[],
 		}