+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Changed user id detection due to a [Rails Port update](https://github.com/openstreetmap/openstreetmap-website/commit/3719e8defbe019b153df79bf6996341d5774759d#diff-78ff736409d758722403cce31873ba803b3fb526d37398ca21caa557b54dfd95)
 - Warn on the settings page that changing settings cancels ongoing actions
+- Don't cancel ongoing actions when closing the settings page
 
 ## 0.1.1